more actualizations and dto clause

+++ src/services/services.js
@@ -1,30 +1,64 @@
 import Client from "../models/Client.js";
+import { ClientDto } from "../dtos/ClientDto.js";
 
 const client = new Client(); 
 
 class clientServices {
     static async list() {
-        return await client.search();
+        const listAll = await client.search();
+
+        return  listAll.map(client => new ClientDto(client));
     }
 
     static async listId(id) {
-        return await client.searchId(id);
+         const listId = await client.searchId(id);
+
+         return new ClientDto(listId);
     }
 
     static async create(id, name, cpf, dateBirth) {
+        const duplicate = await this.verifyCPF(cpf, id);
+    
+        if (duplicate) {
+            throw new Error(`O CPF ${cpf} já existe.`);
+        }
 
-        return await client.create(id, name, cpf, dateBirth);
+        const createdClient = await client.create(id, name, cpf, dateBirth);
+
+        return new ClientDto(createdClient);
     }
 
     static async update(id, name, cpf, dateBirth) {
-        return await client.update(id, name, cpf, dateBirth); 
+        const listClient = await this.listId(id);
+        const duplicate = await this.verifyCPF(cpf, id);
+
+        if(!listClient){
+            console.log("Entrrou")
+            throw new Error(`o ID ${id} não existe.`)
+        }
+
+        if (duplicate) {
+            throw new Error(`o CPF ${cpf} ja existe.`)
+        }
+
+        const updateClient = await client.update(id, name, cpf, dateBirth); 
+
+        return new ClientDto(updateClient);
     }
 
     static async deleteClientId(id){
-        return await client.deleteClient(id);
+        const listClient = await this.listId(id);
+
+        if(!listClient){
+            throw new Error(`o ID ${id} não existe.`)
+        }
+
+        const deleteClient = await client.deleteClient(id);
+
+        return new ClientDto(deleteClient);
     }
 
-    static async verify(cpf, id){
+    static async verifyCPF(cpf, id){
         return await client.verifyCPF(cpf, id); 
     }
 }
